+++ scripts/truffle/script_template.js
