+++ scripts/truffle/write_driip_settlement_state.js
